refactor: rename cf-driver to cfd
